created a separate file for coolest_slsim interface in Util.
